+++ nginx/nginx.conf
@@ -54,7 +54,7 @@ http {
             proxy_set_header Connection 'upgrade';
             proxy_cache_bypass $http_upgrade;
             proxy_set_header SCRIPT_NAME /api;
-            proxy_set_header Host $http_host:$server_port;
+            proxy_set_header Host $http_host;
 
              if ($request_method = 'OPTIONS') {
                 add_header 'Access-Control-Allow-Origin' '*';
@@ -132,14 +132,14 @@ http {
             proxy_redirect off;
         }
 
-        # location /ws/ {
-        #     proxy_pass http://websockets/;
-        #     proxy_http_version 1.1;
-        #     proxy_set_header Upgrade $http_upgrade;
-        #     proxy_set_header Connection 'upgrade';
-        #     proxy_set_header Host $host;
-        #     proxy_cache_bypass $http_upgrade;
-        # }
+        location /ws/ {
+            proxy_pass http://websockets/;
+            proxy_http_version 1.1;
+            proxy_set_header Upgrade $http_upgrade;
+            proxy_set_header Connection 'upgrade';
+            proxy_set_header Host $host;
+            proxy_cache_bypass $http_upgrade;
+        }
 
         location / {
             proxy_pass http://frontend;
