+++ websockets/src/go.mod
@@ -1,0 +1,8 @@
+module marktai.com/websockets
+
+go 1.18
+
+require (
+	github.com/gorilla/mux v1.8.0
+	github.com/gorilla/websocket v1.5.0
+)
