--- websockets/src/main/main.go
@@ -1,16 +1,0 @@
-package main
-
-import (
-	"flag"
-	"server"
-)
-
-func main() {
-	var port int
-
-	flag.IntVar(&port, "port", 80, "Port the server listens to")
-
-	flag.Parse()
-
-	server.Run(port)
-}
